+++ src/app/app.module.ts
@@ -13,7 +13,7 @@ import { MaterialModule } from './material/material.module';
 import { SharedModule } from './shared/shared.module';
 import { AppRoutingModule } from './app-routing/app-routing.module';
 import { CoreModule } from './core/core.module';
-import { NotFoundModule } from './not-found/not-found.module';
+import { NotFoundModule } from './views/not-found/not-found.module';
 import { RequestedLoanModule } from './views/requested-loan/requested-loan.module';
 import { ActiveLoansModule } from './views/active-loans/active-loans.module';
 import { AddressModule } from './views/address/address.module';
