--- src/app/not-found/not-found-routing.module.ts
@@ -1,15 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-// App Component
-import { NotFoundComponent } from './not-found.component';
-
-const routes: Routes = [
-  { path: '404', component: NotFoundComponent }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class NotFoundRoutingModule { }
