--- src/app/not-found/not-found.component.html
@@ -1,16 +1,0 @@
-<div class="margin-spacing" [style.height.px]="winHeight">
-  <div class="txt-container">
-    <div class="title">this loan does not exist</div>
-    <div class="miscellany"></div>
-    <div class="subtitle">Oooops, something went wrong</div>
-  </div>
-  <div class="img-container"></div>
-  <div id="clouds">
-    <div class="cloud x1"></div>
-    <div class="cloud x1_5"></div>
-    <div class="cloud x2"></div>
-    <div class="cloud x3"></div>
-    <div class="cloud x4"></div>
-    <div class="cloud x5"></div>
-  </div>
-</div>
