--- src/app/not-found/not-found.component.scss
@@ -1,83 +1,0 @@
-@import '../../scss/_variables.scss'; //variables
-
-.margin-spacing{
-  position: relative;
-  margin: -20px 0 -60px;
-}
-.txt-container{
-  width: 100%;
-  position: absolute; 
-  bottom: -8px;
-  padding: 15px;
-  text-align: center;
-  background-color: $tertiary;
-  .title{
-    text-transform: uppercase;
-    @extend %font-regular;
-  }
-  .subtitle{
-    font-size: 20px;
-    @extend %font-bold;
-  }
-}
-.img-container{
-  width: 100%;
-  height: 100%;
-  background-image: url(../../assets/404-tree.svg);
-  background-repeat: no-repeat;
-  background-position: center center;
-  @include background-size(cover);
-}
-
-#clouds{position: fixed; top: 0; left: 0; right: 0; bottom: 0; height: 100%; width: 100%; overflow: hidden; z-index: -1;}
-.cloud {width: 350px; height: 120px; background: #FFF; background: linear-gradient(to top, #FFF 100%); background: -webkit-linear-gradient(to top, #FFF 100%); background: -moz-linear-gradient(to top, #FFF 100%); background: -ms-linear-gradient(to top, #FFF 100%); background: -o-linear-gradient(to top, #FFF 100%); border-radius: 100px; -webkit-border-radius: 100px; -moz-border-radius: 100px; position: absolute; margin: 120px auto 20px; z-index:-1; transition: ease 1s; }
-.cloud:after, .cloud:before {content: ''; position: absolute; background: #FFF; z-index: -1 }
-.cloud:after {width: 100px; height: 100px; top: -50px; left: 50px; border-radius: 100px; -webkit-border-radius: 100px; -moz-border-radius: 100px; }
-.cloud:before {width: 180px; height: 180px; top: -90px; right: 50px; border-radius: 200px; -webkit-border-radius: 200px; -moz-border-radius: 200px; }
-.x1 {top:-50px; left:100px; -webkit-transform: scale(0.3); -moz-transform: scale(0.3); transform: scale(0.3); opacity: 0.9; -webkit-animation: moveclouds 15s linear infinite; -moz-animation: moveclouds 15s linear infinite; -o-animation: moveclouds 25s linear infinite; }
-.x1_5{top:-80px; left:250px; -webkit-transform: scale(0.3); -moz-transform: scale(0.3); transform: scale(0.3); -webkit-animation: moveclouds 17s linear infinite; -moz-animation: moveclouds 17s linear infinite; -o-animation: moveclouds 27s linear infinite; }
-.x2 {left: 250px; top:30px; -webkit-transform: scale(0.6); -moz-transform: scale(0.6); transform: scale(0.6); opacity: 0.6; -webkit-animation: moveclouds 25s linear infinite; -moz-animation: moveclouds 25s linear infinite; -o-animation: moveclouds 35s linear infinite; }
-.x3 {left: 250px; bottom: -70px; -webkit-transform: scale(0.6); -moz-transform: scale(0.6); transform: scale(0.6); opacity: 0.8; -webkit-animation: moveclouds 25s linear infinite; -moz-animation: moveclouds 25s linear infinite; -o-animation: moveclouds 35s linear infinite; }
-.x4 {left: 470px; botttom: 20px; -webkit-transform: scale(0.75); -moz-transform: scale(0.75); transform: scale(0.75); opacity: 0.75; -webkit-animation: moveclouds 18s linear infinite; -moz-animation: moveclouds 18s linear infinite; -o-animation: moveclouds 18s linear infinite; }
-.x5 {left: 200px; top: 300px; -webkit-transform: scale(0.5); -moz-transform: scale(0.5); transform: scale(0.5); opacity: 0.8; -webkit-animation: moveclouds 20s linear infinite; -moz-animation: moveclouds 20s linear infinite; -o-animation: moveclouds 30s linear infinite; }
-@-webkit-keyframes moveclouds {0% {margin-left: 2000px;} 100% {margin-left: -500px;} } @-moz-keyframes moveclouds {0% {margin-left: 2000px;} 100% {margin-left: -500px;} } @-o-keyframes moveclouds {0% {margin-left: 2000px;} 100% {margin-left: -500px;} }
-
-@media (min-width: 992px){
-  #clouds{top: 60px; bottom: 58px;}
-  .margin-spacing{
-    margin: -55px 0 -75px;
-    overflow: hidden;
-  }
-  .txt-container{
-    width: 30%;
-    max-width: 500px;
-    top: 50%; 
-    bottom: initial;
-    @include transform(translate(0%, -50%));
-    text-align: left;
-    padding: 0;
-    margin-top: -40px;
-    margin-left: 5%;
-    background-color: transparent;
-    .title{font-size: 20px;}
-    .subtitle{font-size: 35px; line-height: 1;}
-  }
-  .miscellany {
-    width: 40px;
-    border-bottom: 3px solid white;
-    margin: 15px 0 30px;
-  }
-  .img-container{
-    width: 90%;
-    float: right;
-  }
-}
-
-@media (min-width: 1200px){
-  .txt-container{
-    margin-left: 10%;
-    .title{font-size: 30px;}
-    .subtitle{font-size: 50px;}
-  }
-}
-@media (min-width: 1400px){.txt-container{margin-left: 14%;}}
