--- src/app/not-found/not-found.component.ts
@@ -1,16 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-not-found',
-  templateUrl: './not-found.component.html',
-  styleUrls: ['./not-found.component.scss']
-})
-export class NotFoundComponent implements OnInit {
-  winHeight: any = window.innerHeight - 118;
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
